add JWT check to taskRoutes

+++ backend/routes/taskRoutes.js
@@ -82,6 +82,31 @@ taskRouter.patch("/update-task/:id", (request, response) => {
 	console.log("Task ID", request.params.id);
 	console.log("Task data", request.body);
 
+	var id;
+	let sendError = false; // If token can't be verified, or a problem arises.
+
+	try {
+        let token 	= request.cookies["Bearer"]
+		let valid 	= validateJWT( token );
+
+		if (!valid) {
+			sendError = true;
+		}
+
+		id = _id;
+
+    } catch (error) {
+        console.log( error )
+		sendError = true;
+    }   
+
+	if (sendError) {
+		console.log("[!] Could not validate token")
+        response.status(401).send("Invalid token.");
+        return;
+	}
+
+	// TODO: Validate new Task Body somehow
 	taskModel.findByIdAndUpdate(request.params.id, request.body, { new: true })
 		.then((task) => {
 			console.log("[*] Task updated!", task);
@@ -98,6 +123,32 @@ taskRouter.patch("/update-task/:id", (request, response) => {
 taskRouter.delete("/delete-task/:id", (request, response) => {
 	console.log("[>] DELETE '/delete-task/:id'");
 	console.log("Task ID", request.params.id);
+
+	var id;
+	let sendError = false; // If token can't be verified, or a problem arises.
+
+	try {
+        let token 	= request.cookies["Bearer"]
+		let valid 	= validateJWT( token );
+
+		//TODO: Check ownership before deletion !!
+
+		if (!valid) {
+			sendError = true;
+		}
+
+		id = _id;
+
+    } catch (error) {
+        console.log( error )
+		sendError = true;
+    }   
+
+	if (sendError) {
+		console.log("[!] Could not validate token")
+        response.status(401).send("Invalid token.");
+        return;
+	}
 
 	taskModel.findByIdAndDelete(request.params.id)
 		.then((task) => {
